Beat controllers are fully ported.

+++ src/controllers/beatControllers.ts
@@ -1,10 +1,10 @@
 import fs from "fs";
-import Beat from "../src/models/beatModel";
-import Purchase from "../src/models/purchaseModel";
-import Track from "../src/models/trackModel";
-import { APIFeatures } from "./../utils/apiFeatures";
+import Beat, { BeatDocument } from "../models/beatModel";
+import Purchase from "../models/purchaseModel";
+import Track from "../models/trackModel";
 import catchAsync from "./../utils/catchAsync";
 import { Request, Response } from "express";
+import mongoose from "mongoose";
 
 async function getAllBeats(req: Request, res: Response) {
   try {
@@ -26,11 +26,15 @@ async function getAllBeats(req: Request, res: Response) {
 async function createBeat(req: Request, res: Response) {
   try {
     // Photo
-    const newPhoto = req.files.photo[0];
+    const newPhoto = (
+      req.files as { [fieldname: string]: Express.Multer.File[] }
+    ).photo[0];
     const photoPath = newPhoto.path;
 
     // Track
-    const { originalname, path, mimetype } = req.files.fullTrack[0];
+    const { originalname, path, mimetype } = (
+      req.files as { [fieldname: string]: Express.Multer.File[] }
+    ).fullTrack[0];
     let fileType = mimetype.split("/")[1];
     if (fileType === "mpeg") fileType = "mp3";
     const uploader = req.body.owner;
@@ -106,19 +110,27 @@ async function updateBeat(req: Request, res: Response) {
 
 async function deleteBeat(req: Request, res: Response) {
   try {
-    const beat = await Beat.findByIdAndDelete(req.params.id);
-    const tracks = beat.tracks;
-    const fullTrack = await Track.findByIdAndDelete(beat.fullTrack);
-    const photoPath = beat.photo;
-    const fullTrackPath = fullTrack.path;
-    await fs.unlink(photoPath, () => console.log("Photo deleted!"));
-    await fs.unlink(fullTrackPath, () => console.log("FullTrack deleted!"));
+    const beat: BeatDocument | null = await Beat.findByIdAndDelete(
+      req.params.id
+    );
+    const tracks = beat?.tracks || [];
+    const fullTrack = beat
+      ? await Track.findByIdAndDelete(beat.fullTrack)
+      : null;
+    const photoPath = beat?.photo ?? "";
+    const fullTrackPath = fullTrack?.path;
+    fs.unlink(photoPath, () => console.log("Photo deleted!"));
+    if (fullTrackPath) {
+      fs.unlink(fullTrackPath, () => console.log("FullTrack deleted!"));
+    }
 
     if (tracks.length !== 0) {
       for (let i = 0; i < tracks.length; i++) {
         const track = await Track.findByIdAndDelete(tracks[i]);
-        const trackPath = track.path;
-        await fs.unlink(trackPath, () => console.log("Track (i) deleted!"));
+        const trackPath = track?.path;
+        if (trackPath) {
+          fs.unlink(trackPath, () => console.log("Track (i) deleted!"));
+        }
       }
     }
 
@@ -192,7 +204,13 @@ const addTrack = async (req: Request, res: Response) => {
     const { beatId, trackId } = req.params;
 
     const beat = await Beat.findById(beatId);
-    beat.tracks = [...beat.tracks, trackId];
+    if (!beat) {
+      return res.status(404).json({
+        status: "fail",
+        message: "Beat not found",
+      });
+    }
+    beat.tracks = [...beat.tracks, new mongoose.Schema.Types.ObjectId(trackId)];
     const newBeat = await Beat.findByIdAndUpdate(beatId, beat, {
       runValidators: true,
       new: true,
@@ -224,8 +242,14 @@ const deleteTrack = async (req: Request, res: Response) => {
     );
 
     const track = await Track.findByIdAndDelete(trackId);
+    if (!track) {
+      return res.status(404).json({
+        status: "fail",
+        message: "Track not found",
+      });
+    }
     const path = track.path;
-    await fs.unlink(path, () => {
+    fs.unlink(path, () => {
       console.log("Track deleted successfully");
     });
 
@@ -241,12 +265,14 @@ const deleteTrack = async (req: Request, res: Response) => {
   }
 };
 
-const editBeatMain = async (req: Request, res: Response) => {
+const editBeatMain = async (
+  req: Request<{ beatId: string }, unknown, { type: string[] }>,
+  res: Response
+) => {
   try {
-    console.log(req.body);
     const { beatId } = req.params;
 
-    let parsedTypes = [];
+    let parsedTypes: string[] = [];
     if (req.body.type) {
       parsedTypes = req.body.type.map((e) => e.trim());
     }
@@ -281,7 +307,7 @@ const queryBeats = catchAsync(async (req, res) => {
   });
 });
 
-module.exports = {
+export default {
   getAllBeats,
   getBeat,
   createBeat,
